# [WIP] Custom fusedL2NN kernel for kmeans prediction