--- a/cpp/src/distance/fused_l2_nn.cu
+++ b/cpp/src/distance/fused_l2_nn.cu
@@ -52,3 +52,23 @@ instantiate_raft_distance_fusedL2NNMinReduce(float,
 #undef COMMA
 
 #undef instantiate_raft_distance_fusedL2NNMinReduce
+
+
+#define instantiate_raft_distance_fusedL2NNMinReduceCustomKernel(DataT, LabelT, IdxT)                   \
+  template void raft::distance::fusedL2NNMinReduceCustomKernel<DataT, LabelT, IdxT>(LabelT* label,      \
+                                                                                    const DataT* x,     \
+                                                                                    const DataT* y,     \
+                                                                                    const DataT* xn,    \
+                                                                                    const DataT* yn,    \
+                                                                                    IdxT m,             \
+                                                                                    IdxT n,             \
+                                                                                    IdxT k,             \
+                                                                                    bool sqrt,          \
+                                                                                    cudaStream_t stream)
+
+instantiate_raft_distance_fusedL2NNMinReduceCustomKernel(float, uint32_t, int);
+instantiate_raft_distance_fusedL2NNMinReduceCustomKernel(float, uint32_t, int64_t);
+instantiate_raft_distance_fusedL2NNMinReduceCustomKernel(double, uint32_t, int);
+instantiate_raft_distance_fusedL2NNMinReduceCustomKernel(double, uint32_t, int64_t);
+
+#undef instantiate_raft_distance_fusedL2NNMinReduceCustomKernel