This PR adds a custom fusedL2NN kernel in order to speedup PQ codebook generation step of IVF-PQ index building. This PR partially address #1901 (we are also experimenting other ideas, such as subsampling, to optimize the PQ codebook generation step). 

The key kernel to the PQ code generation performance is kmeans predict, which is used to find the cluster that is nearest (defined by distance function) from each data point. Depending on the distance type, this could be viewed as some variants of GEMM between `dataset` and `cluster_centers`. In this PR, we focus on L2 distance. The current implementation utilizes CUTLASS kernel for this L2 distance calculation, with a fused reduction part that performs an `argmin` on the distances and outputs the label of the cluster which is the closest to each input data point.  

In codebook generation step, we have a small input matrix for `cluster_center` (size `[n_cluster, n_dim]`), since `n_dim` depends on `pq_len` which is usually 1-24 and `n_cluster` depends on `2^pq_bit`, which ranges from 16 to 256 (`pq_bit` from [4, 8]). Also, the `dataset` is a very tall matrix (size  `[n_row, n_dim]`), `n_row` is the number of rows in the training set after subsampling, which is usually on the order of millions. 

This PR focus on optimizing the GEMM with thin and small inputs. We found a custom non-GEMM kernel is able to achieve up to 33x compared to the original fused GEMM implementation for this problem when `n_rows` is 10M. For larger `n_clusters` and `n_dim`, the fused GEMM is still the best performer, but for codebook generation part, most of the problem sizes are not ideal for GEMM. Please see the following benchmark results are produced on a A100-80GB-PCIe machine. 

![image](https://github.com/rapidsai/raft/assets/17487689/597d07f3-fa91-4811-b362-975edcba6175)

In order to prevent performance hit, we only use the custom kernel when `n_clusters*n_dim` is less or equal than 256 in this PR. Those cases are circled in green in the picture above. Please let me know if you have ideas on better heuristics. 

Here are E2E results on Deep-100M dataset. More benchmark numbers to be posted. Search performance is not impacted by this PR, Most recall/runtime are within 5% difference; all within 10%. Those variance seems to be run-to-run difference. 

Dataset | n_iter | n_list | pq_bits | pq_dim | ratio | Original time (s) | Optimized time (s) | Speedup
-- | -- | -- | -- | -- | -- | -- | -- | --
Deep-100M | 25 | 50000 | 4 | 96 | 10 | 129 | 105 | 1.23
Deep-100M | 25 | 50000 | 5 | 96 | 10 | 128 | 103 | 1.24
Deep-100M | 25 | 50000 | 6 | 96 | 10 | 131 | 107 | 1.22
Deep-100M | 25 | 50000 | 7 | 96 | 10 | 129 | 107 | 1.21
Deep-100M | 25 | 50000 | 8 | 96 | 10 | 149 | 117 | 1.27

TODO 

- [ ] Unit tests 
- [ ] E2E benchmark numbers (deep-100M, sift-1M)