@abc99lr 2023-12-08
Tagging @tfeher @achirkin @cjnolet for some early review, before running more benchmarks and writing test cases (although I am pretty sure the accuracy should be fine). Thanks 
I'd also like to get some thoughts on 

- which particular cases we should benchmark 
- the heuristic to select when to use the custom kernel. Right now it depends on `n_cluster*n_dim`, if the value is less or equal than 256, use the custom kernel. This is not optimal since it cannot pick up all the cases where the custom kernel could achieve speedup