@abc99lr 2023-12-08
I also created another PR for subsampling support. https://github.com/rapidsai/raft/pull/2052 
With subsampling, this PR doesn't seem to be necessary anymore, but I think it's necessary to justify subsampling won't impact the index quality. Any ideas on what benchmark performance numbers need to be collected?