@abc99lr 2024-01-13
Thanks for the comments @achirkin. However, after discussing with @tfeher, we believe #2052 is a better optimization idea and it cannot work together with this PR. Given that, I think it makes sense to close this PR for now. Will reopen if needed.